+++ tests/test_dcos_e2e/test_node.py
@@ -688,6 +688,14 @@ class TestOutput:
     Tests for the ``output`` parameter of ``Node.run``.
     """
 
+    @pytest.fixture(autouse=True)
+    def configure_logging(self, caplog: LogCaptureFixture) -> None:
+        """
+        Set the ``caplog`` logging level to ``DEBUG`` so it captures any log
+        messages produced by ``dcos_e2e`` library.
+        """
+        caplog.set_level(logging.DEBUG, logger='dcos_e2e')
+
     def test_default(
         self,
         caplog: LogCaptureFixture,